requirements.txt
ultralytics>=8.0.0
opencv-python>=4.6.0
PyQt5>=5.15.0
psutil>=5.9.0
GPUtil>=1.4.0
pandas>=1.3.0
pyyaml>=6.0
openpyxl>=3.0.0
